requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# API interactions
requests>=2.31.0
anthropic>=0.18.0

# Image processing
Pillow>=10.0.0

# Visualization
plotly>=5.17.0
networkx>=3.1

# Development tools
black>=23.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1 